backend/requirements.txt
fastapi
uvicorn
hindsight-client
groq
pydantic
python-dotenv
